[#96] MVVM + UIKit 코드 제거
