+++ Soomsil-USaint/Application/Feature/Semester/SemesterDetail/View/SemesterDetailView.swift
@@ -1,216 +1,212 @@
 //
-//  SemesterDetailView.swift
-//  Soomsil-USaint-iOS 
+//  v2SemesterDetailView.swift
+//  Soomsil-USaint
 //
-//  Created by 최지우 on 12/17/24.
+//  Created by 최지우 on 3/24/25.
 //
 
 import SwiftUI
+
+import ComposableArchitecture
 import YDS_SwiftUI
 
-private enum Dimension {
-    enum Spacing {
-        static let mainVStack = 16.0
-    }
+struct SemesterDetailView: View {
+    @Bindable var store: StoreOf<SemesterDetailReducer>
 
-    enum Size {
-        static let stampImage = 130.0
-        static let entireView = UIScreen.main.bounds.size
-    }
-
-    enum Padding {
-        static let `default` = 16.0
-    }
-
-    enum Origin {
-        static let entireView = UIScreen.main.bounds.origin
-    }
-}
-
-private typealias Spacing = Dimension.Spacing
-private typealias Size = Dimension.Size
-private typealias Padding = Dimension.Padding
-private typealias Origin = Dimension.Origin
-
-struct SemesterDetailView<VM: SemesterDetailViewModel>: View {
-    @Binding var path: [StackView]
-    @Environment(\.dismiss) var dismiss
-    @StateObject var semesterDetailViewModel: VM
-    @State private var viewOrigin: CGPoint = Origin.entireView
-    @State private var viewSize: CGSize = Size.entireView
-    var isShowSummary: Bool = true
-
-    @State private var isLoading = true
+    @State private var mainViewScrollState: SemesterTab.ID?
+    @State private var tabBarScrollState: SemesterTab.ID?
+    @State private var progress: CGFloat = .zero
 
     var body: some View {
-        ZStack {
-            ScrollView {
-                ZStack(alignment: .bottomTrailing) {
-                    VStack(alignment: .leading, spacing: Spacing.mainVStack) {
-                        Text("\(String(semesterDetailViewModel.gradeSummary.year))년 \(semesterDetailViewModel.gradeSummary.semester)")
-                            .font(YDSFont.subtitle2)
-                        if isShowSummary {
-                            HStack(alignment: .lastTextBaseline) {
-                                Text(String(format: "%.2f", semesterDetailViewModel.gradeSummary.gpa))
-                                    .font(YDSFont.display1)
-                                Text("/ 4.50")
-                                    .foregroundColor(YDSColor.textTertiary)
-                            }
-                            GradeOverview(
-                                title: "취득 학점",
-                                accentText: "\(Int(semesterDetailViewModel.gradeSummary.earnedCredit.rounded()))"
-                            )
-                            GradeOverview(
-                                title: "학기별 석차",
-                                accentText: "\(semesterDetailViewModel.gradeSummary.semesterRank)",
-                                subText: "\(semesterDetailViewModel.gradeSummary.semesterStudentCount)"
-                            )
-                            GradeOverview(
-                                title: "전체 석차",
-                                accentText: "\(semesterDetailViewModel.gradeSummary.overallRank)",
-                                subText: "\(semesterDetailViewModel.gradeSummary.overallStudentCount)"
-                            )
-                        }
-                        Divider()
-                        
-                        if let lectures = semesterDetailViewModel.gradeSummary.lectures {
-                            ForEach(Array(lectures.enumerated()), id: \.offset) { index, lecture in
-                                if semesterDetailViewModel.masking {
-                                    MaskedGradeRow(grade: lecture.grade)
-                                } else {
-                                    GradeRow(
-                                        lectureName: lecture.title,
-                                        professor: lecture.professorName,
-                                        credit: lecture.credit,
-                                        grade: lecture.grade
-                                    )
-                                    .offset(x: semesterDetailViewModel.rowAnimation ? 0 : 100)
-                                    .opacity(semesterDetailViewModel.rowAnimation ? 1 : 0)
-                                    .animation(
-                                        .easeIn
-                                            .delay(Double(index) * 0.1)
-                                            .speed(0.5),
-                                        value: semesterDetailViewModel.rowAnimation
-                                    )
-                                    .onAppear {
-                                        withAnimation {
-                                            semesterDetailViewModel.rowAnimation = true
-                                        }
-                                    }
-                                }
-                            }
-                        } else {
-                            HStack {
-                                Spacer()
-                                Text("세부 성적을 불러오려면 새로고침을 터치해주세요.")
-                                Spacer()
-                            }
-                        }
-                    }
-                    .padding(Padding.default)
-                    .background(
-                        GeometryReader { proxy in
-                            Color.clear
-                                .onAppear {
-//                                    adjustScreenshotSize(proxy)
-                                }
-                        }
-                            .padding(.horizontal, -Padding.default) // size 측정 시 버그(약간 축소되어 측정 됨) 때문에 임시방편으로 설정.
-                    )
 
-                    Image("ppussungStamp")
-                        .resizable()
-                        .aspectRatio(1.0, contentMode: .fit)
-                        .frame(maxWidth: Size.stampImage, maxHeight: Size.stampImage)
-                        .padding(.trailing, Padding.default)
-                        .opacity(semesterDetailViewModel.isCapturing ? 1 : 0)
-                }
-            }
-            .background(YDSColor.bgElevated)
-            .toolbar {
-                ToolbarItemGroup(placement: .navigationBarTrailing) {
-                    if #available(iOS 15.0, *) {
-                        Button(action: semesterDetailViewModel.takeScreenshot) {
-                            YDSIcon.cameraLine
-                                .renderingMode(.template)
-                                .foregroundColor(YDSColor.buttonNormal)
-                        }
-                        .confirmationDialog(
-                            "강의 정보를 가리고 저장할까요?",
-                            isPresented: $semesterDetailViewModel.showConfirmDialog,
-                            actions: {
-                                Button("강의 정보 가리고 저장") {
-                                    semesterDetailViewModel.takeScreenshotWithMasking {
-                                        self.screenshot()
-                                    }
-                                }
-                                Button("원본으로 저장") {
-                                    semesterDetailViewModel.takeScreenshotWithoutMasking {
-                                        self.screenshot()
-                                    }
+        VStack(spacing: 0) {
+            TabView(tabs: $store.tabs,
+                    activeTab: $store.activeTab,
+                    mainViewScrollState: $mainViewScrollState,
+                    tabBarScrollState: $tabBarScrollState,
+                    progress: $progress)
+
+            GeometryReader {
+                let size = $0.size
+
+                ScrollView(.horizontal) {
+                    LazyHStack(spacing: 0) {
+
+                        ForEach(store.tabs) { tab in
+                            ScrollView(.vertical) {
+                                let tappedSemester = findTappedSemester(
+                                    semesterList: store.semesterList,
+                                    tabId: tab.id
+                                )
+                                if let semester = tappedSemester {
+                                    TopSummary(gradeSummary: semester)
+                                    GradeList(lectures: semester.lectures ?? [])
                                 }
                             }
-                        )
-                        .alert("저장 완료", isPresented: $semesterDetailViewModel.showSuccessAlert, actions: {})
-                        .alert("저장 실패", isPresented: $semesterDetailViewModel.showFailureAlert, actions: {})
+                            .padding(20)
+                            .frame(width: size.width, height: size.height)
+                            .contentShape(.rect)
+                        }
+                    }
+                    .scrollTargetLayout()
+                    .rect { rect in
+                        progress = -rect.minX / size.width
                     }
                 }
             }
-            .onAppear {
-                Task {
-                    await semesterDetailViewModel.getLectureList()
-                    isLoading = false
-                }
-            }
-            .refreshable {
-                Task {
-                    await semesterDetailViewModel.loadLectureListFromRusaint()
-                }
-            }
-            .onChange(of: semesterDetailViewModel.fetchErrorMessage) {
-                message in
-                YDSToast(message, haptic: .failed)
-            }
-            .registerYDSToast()
-            .navigationBarBackButtonHidden()
-            .toolbar {
-                ToolbarItem(placement: .topBarLeading) {
-                    Button {
-                        dismiss()
-                    } label: {
-                        Image("ic_arrow_left_line")
-                            .resizable()
-                            .frame(width: 24, height: 24)
+            .scrollPosition(id: $mainViewScrollState)
+            .scrollIndicators(.hidden)
+            .scrollTargetBehavior(.paging)
+            .onChange(of: mainViewScrollState) { oldValue, newValue in
+                if let newValue {
+                    debugPrint(newValue)
+                    withAnimation(.snappy) {
+                        tabBarScrollState = newValue
+                        store.activeTab = newValue
                     }
                 }
-            }
-
-            if isLoading {
-                CircleLoadingView()
             }
         }
-    }
-    private func screenshot() {
-        self.screenShot(origin: viewOrigin, size: viewSize) {
-            semesterDetailViewModel.takeScreenshotSuccess()
-        } failure: {
-            semesterDetailViewModel.takeScreenshotFailure()
+        .onAppear() {
+            store.send(.onAppear)
         }
     }
 
-    private func adjustScreenshotSize(_ proxy: GeometryProxy) {
-        DispatchQueue.main.async {
-            if proxy.size != .zero {
-                self.viewOrigin = proxy.frame(in: SwiftUI.CoordinateSpace.local).origin
-                self.viewSize = proxy.size
+    struct TopSummary: View {
+        var gradeSummary: GradeSummary
+
+        var body: some View {
+            VStack(alignment: .leading) {
+                HStack(alignment: .lastTextBaseline) {
+                    Text(String(format: "%.2f", gradeSummary.gpa))
+                        .font(YDSFont.display1)
+                    Text("/ 4.50")
+                        .foregroundColor(YDSColor.textTertiary)
+                }
+                GradeOverView(title: "취득 학점",
+                                accentText: "\(gradeSummary.earnedCredit)")
+                GradeOverView(title: "학기별 석차",
+                                accentText: "\(gradeSummary.semesterRank)",
+                                subText: "\(gradeSummary.semesterStudentCount)")
+                GradeOverView(title: "전체 석차",
+                                accentText: "\(gradeSummary.overallRank)",
+                                subText: "\(gradeSummary.overallStudentCount)")
+                Divider()
             }
         }
+    }
+
+    struct GradeList: View {
+        var lectures: [LectureDetail]
+
+        var body: some View {
+            VStack {
+                ForEach(lectures, id: \.self.code) { lecture in
+                    GradeRowView(lectureDetail: lecture)
+                }
+            }
+        }
+    }
+
+    private func findTappedSemester(semesterList: [GradeSummary], tabId: String) -> GradeSummary? {
+        let tappedSemesterList = store.semesterList.filter { list in
+            let id = "\(list.year)년 \(list.semester)"
+            return id == tabId
+        }
+        return tappedSemesterList.first
     }
 }
 
-//struct ReportDetailView_Previews: PreviewProvider {
-////    @State private var report = GradeSummaryModel([:])!
-//    static var previews: some View {
-//        SemesterDetailView<SemesterDetailViewModel>(semesterDetailViewModel: SemesterDetailViewModel())
-//    }
-//}
+extension SemesterDetailView {
+
+    struct TabView: View {
+        @Binding var tabs: [SemesterTab]
+        @Binding var activeTab: SemesterTab.ID
+        @Binding var mainViewScrollState: SemesterTab.ID?
+        @Binding var tabBarScrollState: SemesterTab.ID?
+        @Binding var progress: CGFloat
+
+        var body: some View {
+            ScrollView(.horizontal) {
+                HStack(spacing: 20) {
+                    ForEach($tabs) { $tab in
+                        Button(action: {
+                            withAnimation(.snappy) {
+                                activeTab = tab.id
+                                tabBarScrollState = tab.id
+                                mainViewScrollState = tab.id
+                            }
+                        }) {
+                            Text(formatShortedYear(tab.id))
+                                .font(YDSFont.button2)
+                                .padding(.vertical, 12)
+                                .foregroundStyle(activeTab == tab.id ? .primary : Color.gray)
+                                .contentShape(.rect)
+                        }
+                        .buttonStyle(.plain)
+                        .rect { rect in
+                            tab.size = rect.size
+                            tab.minX = rect.minX
+                        }
+                    }
+                }
+                .scrollTargetLayout()
+            }
+            .scrollPosition(id: .init(get: {
+                return tabBarScrollState
+            }, set: { _ in
+
+            }), anchor: .center)
+
+            .overlay(alignment: .bottom) {
+                ZStack(alignment: .leading) {
+                    Rectangle()
+                        .fill(.gray.opacity(0.3))
+                        .frame(height: 2)
+
+                    let inputRange = tabs.indices.compactMap {
+                        return CGFloat($0)
+                    }
+                    let outputRange = tabs.compactMap {
+                        return $0.size.width
+                    }
+                    let outputPositionRange = tabs.compactMap {
+                        return $0.minX
+                    }
+
+                    if !(inputRange.isEmpty || outputRange.isEmpty || outputPositionRange.isEmpty) {
+                        let indicatorWidth = progress.interpolate(inputRange: inputRange, outputRange: outputRange)
+                        let indicatorPosition = progress.interpolate(inputRange: inputRange, outputRange: outputPositionRange)
+
+                        Rectangle()
+                            .fill(.primary)
+                            .frame(width: indicatorWidth,height: 1.5)
+                            .offset(x: indicatorPosition)
+                    }
+                }
+            }
+            .safeAreaPadding(.horizontal, 15)
+            .scrollIndicators(.hidden)
+        }
+
+        private func formatShortedYear(_ id: String) -> String {
+            let components = id.split(separator: "년")
+            guard let year = components.first, components.count > 1 else {
+                return id
+            }
+            if let year = Int(year) {
+                let shortedYear = year % 100
+                return "\(shortedYear)년\(components[1])"
+            }
+            return id
+        }
+    }
+
+}
+
+#Preview {
+    let store = Store(initialState: SemesterDetailReducer.State()) {
+        SemesterDetailReducer()
+    }
+    SemesterDetailView(store: store)
+}
