--- Soomsil-USaint/Global/Utils/Base/Coordinator.swift
@@ -1,31 +1,0 @@
-//
-//  Coordinator.swift
-//  Rusaint-iOS
-//
-//  Created by 이조은 on 12/16/24.
-//
-
-import UIKit
-
-protocol Coordinator: AnyObject {
-    var childCoordinators: [Coordinator] { get set }
-    func start()
-    func start(childCoordinator: Coordinator)
-    func didFinish(childCoordinator: Coordinator)
-    func removeChildCoordinators()
-}
-
-extension Coordinator {
-    public func addChildCoordinator(_ childCoordinator: Coordinator) {
-        self.childCoordinators.append(childCoordinator)
-    }
-
-    public func removeChildCoordinator(_ childCoordinator: Coordinator) {
-        self.childCoordinators = self.childCoordinators.filter { $0 !== childCoordinator }
-    }
-
-    public func removeAllChildCoordinators() {
-        childCoordinators.forEach { $0.removeChildCoordinators() }
-        childCoordinators.removeAll()
-    }
-}
