--- Soomsil-USaint/Global/Utils/Toast/ToastDelegate.swift
@@ -1,12 +1,0 @@
-//
-//  ToastDelegate.swift
-//  Keyboard
-//
-//  Created by Gyuni on 2021/11/15.
-//
-
-import Foundation
-
-protocol ToastDelegate: AnyObject {
-    func showToast()
-}
